yuv_utils: fixup dirty blocknum() parameters
This needs to maintain dirty blocks in the yuv frame coordinates,
not in the captured image's cooridinates.

The existing code assumes those are the same in double-buffer mode
where the dirty block tracking is applied.  But I'm working towards
allowing the captured image dimensions to be free from the 16x16
alignment constraints the yuv buffers must conform to for theora.

To handle that future, the block coordinates should at least use
the (yuv)->y_width, not width_tm.  I also added shifting the x,y
but in double-buffer mode that shouldn't actually be needed since
it'd always be full-shots unless I'm mistaken.