rectinsert: stop aligning rects on insert here
Assume rects that come in for insertion are already as aligned as
possible within the rrect bounds.  If the rrect has odd dimensions,
then there's potential for edge case odd rects too - but the only
even-sensitive code is the YUV updating and that's been amended to
at least ignore those edge cases gracefully.

Also constify the supplied xrect while in here.