get_frame: don't let avd influence frame sampling
Maybe this made sense at some point in the original code, but the
way I have this setup currently get_frame() should strictly
capture a frame on every tick of the timer at the desired FPS to
the best of its ability.

The capture_frameno gets propagated to the encoder whenever a new
frame is acquired on that timer.  When the encoder consumes it,
it should just dupe the frame to fill any gaps between the last
encoded frameno and the new one.

As-is, this avd value seems to drift permanently negative
eventually at which point get_frame() ceases ever waiting on the
timer.  That's obviously broken, and devolves into a pinned CPU
with get_frame() attempting an infinitely high frame rate.  Which
likely just makes things worse not better by starving the encoder
of CPU time.

I need to go check out the encoder now to make sure it fills
frameno gaps.