cache: pivot to new CacheFile API
This minimally switches all the ad-hoc image cache files handling
over to using CacheFile.

I left the audio cache alone for now as it seems to not be
compressed.  It might make sense in the future to switch that
over as well, especially if I start adding features to CacheFile
like preallocating and async periodic fdatasync.