cache: overload syncer thread to also pre-allocate
This changes the syncer thread to always be created for writers,
and makes it posix_fallocate() the cache file to the maximum
size.

When the syncer is disabled w/ms=0, it simply returns after the
posix_fallocate.

posix_fallocate errors are ignored, this is simply an
opportunistic optimization.  I have some concern about the glibc
emulation mentioned in the man page, but writers are opened
O_WRONLY which the notes say cause the emulation to fail with
EBADF which is preferable, and I'm using O_WRONLY for the
writers.