cache: _rmdCacheFileClose should always NULL handles
The existing code would leave the internal handle non-NULL when
the underlying close failed.  In the event of a transparent close
when chapters switch failing, a subsequent explicit close is
expected, and the handles must be NULLed out for that explicit
close to not blow up because fclose() leaves the handle undefined
after errors.  It's basically a use after free bug.