get_frames: remove superfluous avd assignment
Since avd is unconditionally copied just a few lines down,
this copy is pointless and harms readability.

I think when it was originally written there was going to be an
else branch where the unconditional copy is currently, but never
bothered.

The avd shared variable should really get changed to an atomic,
removing its lock.  To anyone reading: PRs welcome.