get_frames: nix even coord constraint in rmdMoveCaptureArea
We no longer bother with such things in this fork.  It's worth
noting the original codebase bothered with these things in places
where it was altogether unnecessary.

This means --follow-mouse will precisely follow the mouse, not
only when the mouse has moved out of the bounding area enough to
reach the next even coordinate.