parseargs,error: add --need-shortcuts
When you depend on the shortcuts as _the_ means of
stopping/pausing a recording, like in a WM integration scenario,
it's preferable to have rmd exit fatally when the grabs can't be
setup.  (usually this occurs when you've accidentally tried
starting multiple recordings, and the shortcuts collide)

This way, the integration can detect the failure and throw up a
dialog or something informing the user that the recording wasn't
started.  All that's needed is including --need-shortcuts.

Note this is a little janky in how it's implemented, because
XSetErrorHandler() is janky and doesn't even support a user
pointer payload.

Rather than introducing global variables to either communicate
the grab failures to other parts of rmd, or exposing the parsed
args struct to the error handler globally in the other direction,
I just made a grab-errors-are-fatal variant of the handler which
gets installed when --need-shortcuts is specified.